Implemented the auth class
