+++ app/Http/Controllers/Auth/LoginController.php
@@ -2,59 +2,11 @@
 
 namespace App\Http\Controllers\Auth;
 
-use App\Http\Controllers\ApiController;
 use App\Http\Controllers\Controller;
-use App\User;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 
-class LoginController extends ApiController
+class LoginController extends Controller
 {
-
-    public function __construct()
-    {
-
-    }
-
-
-    public function login(Request $request)
-    {
-        $this->validate($request, [
-            'email' => 'required|string',
-            'password' => 'required|string'
-        ]);
-
-        $client = DB::table('oauth_clients')
-        ->where('password_client', true)
-        ->first();
-
-        $data = [
-            'grant_type' => 'password',
-            'client_id' => $client->id,
-            'client_secret' => $client->secret,
-            'username' => $request->email,
-            'password' => $request->password
-        ];
-
-        $requestToken = Request::create('/oauth/token', 'POST', $data);
-        $content = json_decode(app()->handle($requestToken)->getContent());
-
-        // dd($content);
-
-        $user = User::where('email', $request->email)->first();
-
-        return $this->showData([
-            'user' => $user,
-            'token' => $content->access_token,
-            'refresh_token' => $content->refresh_token
-        ]);
-        
-        
-    }
-
-     
-
     /*
     |--------------------------------------------------------------------------
     | Login Controller
@@ -66,24 +18,22 @@ class LoginController extends ApiController
     |
     */
 
-    // use AuthenticatesUsers;
+    use AuthenticatesUsers;
 
     /**
      * Where to redirect users after login.
      *
      * @var string
      */
-    // protected $redirectTo = '/home';
+    protected $redirectTo = '/home';
 
     /**
      * Create a new controller instance.
      *
      * @return void
      */
-    // public function __construct()
-    // {
-    //     $this->middleware('guest')->except('logout');
-    // }
-
-
+    public function __construct()
+    {
+        $this->middleware('guest')->except('logout');
+    }
 }
